requirements.txt
SpeechRecognition
translate
monsterapi  
googletrans
pillow
PyAudio
pydantic
langdetect